+++ app/models/api_access_token.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class ApiAccessToken < ApplicationRecord
+  include Uuidable
+
   encrypts :value, deterministic: true
 
   belongs_to :user