+++ app/serializers/api_access_token_serializer.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class ApiAccessTokenSerializer < ApplicationSerializer
+  attributes :uuid, :value
+end