+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - co-owners for accounts
 - expiration time for access tokens
+- creating/removing API access tokens
 
 ### Modified
 - skip reseting invites email after accepting invite